refactor(gui): decompose ConsoleInfoPanel and adopt readout telemetry
